EverythingTogether/requirements.txt
Babel>=2.9.1
Pillow>=8.3.1
pytz>=2021.1
tkcalendar>=1.6.1
pyftpdlib